# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise HRBP Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in as a dev persona
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Invalid persona
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /attrition:
    get:
      operationId: listAttrition
      tags: [attrition]
      summary: List attrition records in scope
      parameters:
        - name: businessUnitId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Attrition records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attrition"
    post:
      operationId: createAttrition
      tags: [attrition]
      summary: Create an attrition record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttritionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attrition"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /attrition/{id}:
    get:
      operationId: getAttrition
      tags: [attrition]
      summary: Get an attrition record
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Attrition record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attrition"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteAttrition
      tags: [attrition]
      summary: Delete an attrition record
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /dashboard/kpis:
    get:
      operationId: getDashboardKpis
      tags: [dashboard]
      summary: Get scope-aware KPIs
      parameters:
        - name: businessUnitId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: KPIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kpis"
  /dashboard/comparison:
    get:
      operationId: getDashboardComparison
      tags: [dashboard]
      summary: Compare KPIs across business units (CHRO/Director/Admin)
      responses:
        "200":
          description: Per business unit KPI rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComparisonRow"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /dashboard/trends:
    get:
      operationId: getDashboardTrends
      tags: [dashboard]
      summary: Get monthly headcount and attrition trend
      parameters:
        - name: businessUnitId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Trend points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List users (Admin only)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a user (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/users/{id}:
    get:
      operationId: getUser
      tags: [admin]
      summary: Get a user (Admin only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Update a user (Admin only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user (Admin only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginInput:
      type: object
      required: [persona]
      properties:
        persona:
          type: string
    AuthUser:
      type: object
      required: [id, email, displayName, role]
      properties:
        id:
          type: integer
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
        businessUnitId:
          type: ["integer", "null"]
        businessUnitName:
          type: ["string", "null"]
...
    Attrition:
      type: object
      required: [id, employeeId, businessUnitId, attritionType, exitDate, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        businessUnitId:
          type: integer
        businessUnitName:
          type: ["string", "null"]
        attritionType:
          type: string
          enum: [voluntary, involuntary]
        reason:
          type: ["string", "null"]
        exitDate:
          type: string
        createdAt:
          type: string
    AttritionInput:
      type: object
      required: [employeeId, businessUnitId, attritionType, exitDate]
      properties:
        employeeId:
          type: integer
        businessUnitId:
          type: integer
        attritionType:
          type: string
          enum: [voluntary, involuntary]
        reason:
          type: ["string", "null"]
        exitDate:
          type: string
...
    User:
      type: object
      required: [id, email, displayName, role, isActive, createdAt, updatedAt]
      properties:
        id:
          type: integer
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [HRBP, HR_DIRECTOR, CHRO, ADMIN]
        businessUnitId:
          type: ["integer", "null"]
        businessUnitName:
          type: ["string", "null"]
        entraObjectId:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    UserInput:
      type: object
      required: [email, displayName, role]
      properties:
        email:
          type: string
          minLength: 1
        displayName:
          type: string
          minLength: 1
        role:
          type: string
          enum: [HRBP, HR_DIRECTOR, CHRO, ADMIN]
        businessUnitId:
          type: ["integer", "null"]
        isActive:
          type: boolean
    UserUpdate:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        displayName:
          type: string
          minLength: 1
        role:
          type: string
          enum: [HRBP, HR_DIRECTOR, CHRO, ADMIN]
        businessUnitId:
          type: ["integer", "null"]
        isActive:
          type: boolean
...
    Kpis:
      type: object
      required:
        [headcount, saudizationPct, femalePct, openRoles, attritionPct, openErCases]
      properties:
        headcount:
          type: integer
        saudizationPct:
          type: number
        femalePct:
          type: number
        openRoles:
          type: integer
        attritionPct:
          type: number
        openErCases:
          type: integer
    ComparisonRow:
      type: object
      required:
        [
          businessUnitId,
          businessUnitName,
          headcount,
          saudizationPct,
          femalePct,
          openRoles,
          attritionPct,
          openErCases,
        ]
      properties:
        businessUnitId:
          type: integer
        businessUnitName:
          type: string
        headcount:
          type: integer
        saudizationPct:
          type: number
        femalePct:
          type: number
        openRoles:
          type: integer
        attritionPct:
          type: number
        openErCases:
          type: integer
    TrendPoint:
      type: object
      required: [month, headcount, exits]
      properties:
        month:
          type: string
        headcount:
          type: integer
        exits:
          type: integer
...