# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise HRBP Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  email: string;
  displayName: string;
  role: UserRole;
  /** @nullable */
  businessUnitId?: number | null;
  /** @nullable */
  businessUnitName?: string | null;
  /** @nullable */
  entraObjectId?: string | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}